refactor: packages/example
fix: types

fix: packages/example

next-env.d.ts

fix: next/navigation-types/compat/navigation
