+++ packages/example/src/types/next-auth.d.ts
@@ -2,7 +2,7 @@ import type { DefaultSession } from 'next-auth';
 
 declare module 'next-auth' {
   /**
-   * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
+   * Returned by `useSession`, `auth`, and received as a prop on the `SessionProvider` React Context
    */
   interface Session {
     /** The user's public Ethereum address. */
